requirements.txt
# Use Command: pip3 install -r requirements.txt
# to install dependencies
PyCryptodome>=3.4.7
# hashlib --Part of base python lib
# Tkinter==2.7 --Part of the base python lib
splinter==0.7.7
